DSPBRIDGE: pm: use pm APIs for opp list
Use the new apis for generating the opp list for DSP.

Original patch: http://marc.info/?l=linux-omap&m=127189549422987&w=2

[fg: ported to L23I3]
[orl: ported to L25]

+++ drivers/dsp/bridge/wmd/tiomap3430.c
@@ -84,7 +84,7 @@
 
 #define MMU_GFLUSH 0x60
 
-extern unsigned short min_active_opp;
+extern unsigned long min_active_opp;
 
 /* Forward Declarations: */
 static DSP_STATUS WMD_BRD_Monitor(struct WMD_DEV_CONTEXT *pDevContext);
